pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "testboiler"
version = "0.1.0"
description = "Boilerplate for quickly testing Python libraries with pytest and unittest"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
testboiler = "testboiler.__main__:main"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.data-files]
"share/testboiler/template" = ["template/quickboiler.cfg", "template/requirements.txt"]
"share/testboiler/template/tests/pytest" = [
    "template/tests/pytest/test_1.py",
    "template/tests/pytest/test_with_file.py",
]
"share/testboiler/template/tests/unittest" = [
    "template/tests/unittest/test_1.py",
    "template/tests/unittest/test_with_file.py",
]
